move loadStoredContentToUI function to newly created popup.utils.js
